refactor(cvtservice): mirror test file split
Move tests into the same files as their corresponding source post-split.
- fixture_generator_test.go: TestGenerateFixture* (5 tests)
- producer_validation_test.go: TestValidateProducerResponse* (8 tests)
- consumer_registry_test.go: Test{Register,List,Deregister}Consumer
- deployment_safety_test.go: TestCanIDeploy*, TestFilterChangesForConsumer

Final part of the validator_service.go split (issue #107).
validator_service_test.go now 1076 LOC (was 2068).
